src/requirements.txt
requests
PyQt5
pillow
